ieiprio: Priorizacion de Variantes en Genes de Errores Innatos de la Inmunidad

Lee archivos VCF, filtra variantes en genes de errores innatos de la inmunidad segun la clasificacion IUIS, las anota con Ensembl VEP y las prioriza segun modo de herencia y fenotipo clinico, generando un reporte por paciente. Herramienta de apoyo e investigacion, no de diagnostico.

See also

Author

Maintainer: Tu Nombre tu.correo@ejemplo.com

Authors:

Tu Nombre tu.correo@ejemplo.com